+++ src/au/com/jwatmuff/eventmanager/gui/results/ResultsPointsPanel.java
@@ -126,11 +126,11 @@ public class ResultsPointsPanel extends javax.swing.JPanel implements Transactio
         }
         Collections.sort(ri, new Comparator<ResultInfo>() {
             public int compare(ResultInfo ri1, ResultInfo ri2) {
-                int[] scores1 = ri1.getResult().getSimpleScores(database);
+                double[] scores1 = ri1.getResult().getSimpleScores(database);
                 int winnerIndex1 = scores1[1] > scores1[0] ? 1 : 0;
                 Player winner1 = ri1.getPlayer()[winnerIndex1];
 
-                int[] scores2 = ri2.getResult().getSimpleScores(database);
+                double[] scores2 = ri2.getResult().getSimpleScores(database);
                 int winnerIndex2 = scores2[1] > scores2[0] ? 1 : 0;
                 Player winner2 = ri2.getPlayer()[winnerIndex2];
                 return winner1.getVisibleID().compareTo(winner2.getVisibleID());
@@ -160,7 +160,7 @@ public class ResultsPointsPanel extends javax.swing.JPanel implements Transactio
                     Date censusDate = database.get(CompetitionInfo.class, null).getAgeThresholdDate();
                     ConfigurationFile configurationFile = ConfigurationFile.getConfiguration(database.get(CompetitionInfo.class, null).getDrawConfiguration());
 
-                    int[] scores = bean.getResult().getSimpleScores(database);
+                    double[] scores = bean.getResult().getSimpleScores(database);
 
                     Grade winnerGrade;
                     Grade loserGrade = null;
@@ -214,7 +214,7 @@ public class ResultsPointsPanel extends javax.swing.JPanel implements Transactio
         private int calculatePoints(ResultInfo info, Grade loserGrade) {
             if(loserGrade == null) return 0;
 
-            int[] score = info.getResult().getSimpleScores(database);
+            double[] score = info.getResult().getSimpleScores(database);
             if(score[0] == score[1]) return 0;
             int w = (score[0] > score[1]) ? 0 : 1;
             int l = 1-w;
